add frontend code to add player to db
